# openapi.yaml
openapi: 3.1.0
...
info:
  title: Hoply API
  version: 1.0.0
  description: |
    Hoply connects travellers heading in the same direction so they can share a
    taxi. Users publish a **travel plan** (source, destination, time window,
    luggage), discover compatible travellers, and exchange **match** requests.
    Contact details are revealed only after both parties accept.

    ### Conventions
    - All timestamps are ISO 8601 / RFC 3339 in UTC (e.g. `2026-07-30T09:15:00Z`).
    - IDs are opaque strings (server generates them).
    - Errors follow [RFC 9457 Problem Details](https://www.rfc-editor.org/rfc/rfc9457).
    - Money is returned in minor units (paise) with an ISO 4217 `currency`.
  contact:
    name: Hoply Engineering
    url: https://hoply.example.com
    email: api@hoply.example.com
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://api.hoply.example.com/v1
    description: Production
  - url: https://staging-api.hoply.example.com/v1
    description: Staging
  - url: http://localhost:8080/v1
    description: Local development
...
security:
  - bearerAuth: []

# =============================================================================
# Paths
# =============================================================================
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      operationId: registerUser
      security: []          # public endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserRegistration' }
      responses:
        '201':
          description: User created.
          headers:
            Location:
              description: URI of the newly created user.
              schema: { type: string, format: uri }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /auth/token:
    post:
      tags: [Auth]
      summary: Exchange credentials for an access token
      operationId: issueToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Token issued.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ----------------------------------------------------------------- Users ---
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Stateless JWT access token in the `Authorization` header.

  # ------------------------------------------------------------- Parameters ---
...
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    Conflict:
      description: Conflicts with current state (e.g. duplicate).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    UnprocessableEntity:
      description: Semantic validation failed.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }

  # ----------------------------------------------------------------- Schemas ---
  schemas:
...
    Gender:
      type: string
      enum: [male, female, other, prefer_not_to_say]
...
    UserRegistration:
      type: object
      required: [name, email, password, phoneNumber]
      properties:
        name:        { type: string, minLength: 1, maxLength: 100 }
        email:       { type: string, format: email, maxLength: 255 }
        password:    { type: string, format: password, minLength: 8, maxLength: 128 }
        phoneNumber: { type: string, maxLength: 20, example: "+919876543210" }
        age:         { type: integer, minimum: 13, maximum: 120, nullable: true }
        gender:      { $ref: '#/components/schemas/Gender' }
        description: { type: string, maxLength: 1000, nullable: true }
...
    User:
      type: object
      properties:
        userId:       { type: string, readOnly: true }
        name:         { type: string }
        email:        { type: string, format: email }
        phoneNumber:  { type: string }
        age:          { type: integer, nullable: true }
        gender:       { $ref: '#/components/schemas/Gender' }
        description:  { type: string, nullable: true }
        profilePhoto: { type: string, format: uri, nullable: true }
        createdAt:    { type: string, format: date-time, readOnly: true }
        updatedAt:    { type: string, format: date-time, readOnly: true }
      required: [userId, name, email, phoneNumber, gender]
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:    { type: string, format: email }
        password: { type: string, format: password }
    TokenResponse:
      type: object
      properties:
        accessToken: { type: string }
        tokenType:   { type: string, example: Bearer }
        expiresIn:   { type: integer, description: Lifetime in seconds., example: 3600 }
      required: [accessToken, tokenType, expiresIn]

    # ---- Locations ----
...
    Problem:
      type: object
      description: Error response per RFC 9457 (Problem Details for HTTP APIs).
      properties:
        type:     { type: string, format: uri, default: about:blank }
        title:    { type: string, example: Validation failed }
        status:   { type: integer, example: 422 }
        detail:   { type: string, example: departUntil must be after departFrom. }
        instance: { type: string, format: uri }
        errors:
          type: array
          description: Field-level validation errors.
          items:
            type: object
            properties:
              field:  { type: string }
              detail: { type: string }
      required: [title, status]